reverse engeniere

+++ AHM_TOTAL_TRAVEL_API/AHM.Total.Travel.Common/Models/HabitacionesViewModel.cs
@@ -13,7 +13,7 @@ namespace AHM.Total.Travel.Common.Models
         public int? Hote_ID { get; set; }
         public decimal? Habi_Precio { get; set; }
         public int? Habi_capacidad { get; set; }
-        public byte? Habi_camas { get; set; }
+        public int? Habi_camas { get; set; }
         public byte? Habi_wifi { get; set; }
         public int? Habi_balcon { get; set; }
         public string Habi_url { get; set; }
